pyproject.toml
[project]
name = "legal-research-assistant-mcp"
version = "0.1.0"
description = "Advanced legal research MCP with treatment analysis, citation networks, and quote verification"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.2",
    "tenacity>=8.2.3",
    "typing-extensions>=4.9.0",
    "pytest-mock>=3.15.1",
    "sentence-transformers>=5.1.2",
    "chromadb>=1.3.5",
    "diskcache>=5.6.3",
    "tenacity>=9.0.0",
]

[project.scripts]
legal-research-mcp = "app.server:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.15.1",
    "respx>=0.21.1",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
allow_untyped_decorators = true

[[tool.mypy.overrides]]
module = ["fastmcp", "fastmcp.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast-running unit tests",
    "integration: slower integration or end-to-end tests",
]
